app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
tabula-py==2.9.0

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web scraping
requests==2.31.0
aiohttp==3.9.1

# ML/Stats
scikit-learn==1.3.2
scipy==1.11.4

# Geospatial
geopandas==0.14.1
shapely==2.0.2

# Network analysis
networkx==3.2.1

# Other utilities
python-dotenv==1.0.0
python-multipart==0.0.6